Fix Mi Flora (#40)
* Fix Mi Flora

Extract missing gatttool which is needed by the Mi Flora component from the edge package.. This is a temporary fix until Alpine 3.7 is released.

* Update Dockerfile

* Include Architectures when downloading gatttool

Not quite elegant, but should do the job.

* Code Cleanup

This sould be more clean than the version before. 
I check $APK_ARCH again just in case. not sure if this is really necessary.

* Update Dockerfile