Fix test for local docker socket
The `-f` test checks if a given file is a "regular file". `docker.sock` will never be a regular file, since it is a socket! The "most correct" test for this would be `-S`, but `-e` is even looser, and I can't see a downside to using `-e`

Documentation:

Proof of concept:

```
$ ls /var/run/docker.sock
/var/run/docker.sock
$ if [ -f /var/run/docker.sock ]; then echo "test passed"; fi
$ if [ -S /var/run/docker.sock ]; then echo "test passed"; fi
test passed
$ if [ -e /var/run/docker.sock ]; then echo "test passed"; fi
test passed
```