Add generic platforms
When setting up a machine for the first time, it took me quite a while
to figure out that the qemu images are suitable as a base image for
my unsupported platform.

This patch adds generic platforms that can either be used outright,
or as a template for a new platform.

Signed-off-by: Alastair D'Silva <alastair@d-silva.org>